+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Update tested platforms
 - Standardize helper library and clean it up
 - Cleanup and modernize unit tests
+- Fix up generic installation
 
 ## 3.0.2 - *2022-12-13*